refactor: 게시글 조회 좋아요 정보 표시 리펙토링
